IMPALA-12904: test_type_conversions_hive3 silently passes because of wrongly defined test dimensions
test_type_conversions_hive3 silently passes because we are not creating
the test dimenstion for query option orc_shema_resolution correctly. If
we set orc_shema_resolution correctly, i.e. to also exercise the
name-based schema resolution, the test fails. The cause of the failure
is that the ill-typed tables have dummy column names like 'c1', 'c2',
etc. These are completely fine for position-based schema resolution,
but it is not OK for name-based schema resolution.

The test just wants to check error messages related to type errors,
the column names are irrelevant, so we can just use the correct
names.

Change-Id: I786a5eaae9243b4728484f3f3b1427b20a1d2d28
Reviewed-on: http://gerrit.cloudera.org:8080/21151
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>