IMPALA-12896 (Part 2): JDBC table must be created as external table
In some of the deployment environments, default table type is
transactional. In these scenarios, JDBC tables which are created as non
external table are not accepted by HMS due to strict managed table check
failures.

This patch forces JDBC tables to be created as external table, and
requires at least 1 column for JDBC tables.

Testing:
 - Updated frontend unit tests and end-to-end unit tests to create JDBC
   tables as external tables.
 - Passed core tests

Change-Id: Ib5533b52434cdf1c430e30ac28a0146ab4d9d4b9
Reviewed-on: http://gerrit.cloudera.org:8080/21159
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>