IMPALA-12923: Fix header alignment during horizontal scrolling in query timeline
On horizontally scrolling the query timeline after zooming, the controls
and navigation bar at the top of the page move away from the visibility
area. This makes it difficult to use the controls while scrolling.

The controls and navigation bar now stay within the visibility area even
after scrolling horizontally, allowing for easier navigation.

Hovering on the edges to scroll horizontally is also supported as before.

Change-Id: Ib3b248554ba32766c29faf91c791f2dbbd20641e
Reviewed-on: http://gerrit.cloudera.org:8080/21173
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>