IMPALA-12612: SELECT * queries expand complex type columns from Iceberg metadata tables
Similarly to how regular tables behave, the nested columns are omitted
when we do a SELECT * on Iceberg metadata tables and the user needs to
turn EXPAND_COMPLEX_TYPES on to include the nested columns into the
result. This patch changes this behaviour to unconditionally include
the nested columns from Iceberg metadata tables.
Note, the behavior of handling nested columns from regular tables
doesn't change with this patch.

Testing:
  - Adjusted the SELECT * metadata table queries to add the nested
    columns into the results.
  - Added some new tests where both metadata tables and regular tables
    were queried in the same query.

Change-Id: Ia298705ba54411cc439e99d5cb27184093541f02
Reviewed-on: http://gerrit.cloudera.org:8080/21236
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>